Caesar Cipher
This program will encode or decode your massage according to ceaser cipher. For simple clarification, there are two types of alphabets, when you entered your shift number one of the other alphabets move as much as that number and that is how all the new letters in your massage is created.